Fix SectionShape circular dependency
