+++ lavamoat/build-system/policy.json
@@ -2484,7 +2484,8 @@
         "eslint-plugin-jest>@typescript-eslint/experimental-utils>@typescript-eslint/types": true,
         "eslint-plugin-jest>@typescript-eslint/experimental-utils>eslint-utils": true,
         "eslint>eslint-scope": true,
-        "eslint>eslint-utils": true
+        "eslint>eslint-utils": true,
+        "webpack>eslint-scope": true
       }
     },
     "eslint-plugin-jest>@typescript-eslint/experimental-utils>eslint-utils": {
@@ -2505,6 +2506,7 @@
         "eslint-plugin-jest>@typescript-eslint/utils>@typescript-eslint/scope-manager": true,
         "eslint-plugin-jest>@typescript-eslint/utils>@typescript-eslint/types": true,
         "eslint-plugin-jest>@typescript-eslint/utils>eslint-utils": true,
+        "eslint-plugin-jest>@typescript-eslint/utils>webpack>eslint-scope": true,
         "eslint>eslint-scope": true,
         "eslint>eslint-utils": true,
         "webpack>eslint-scope": true
