+++ ui/components/app/metamask-template-renderer/index.js
@@ -1,1 +1,2 @@
 export { default } from './metamask-template-renderer';
+export { SectionShape } from './section-shape';
